Update README.md
Added a Personal Point of View of the Course.

As every repository that I'm creating, my intention is to be learning and be more familiar with the software (In this case, the main sector is Dart & Flutter, but trying to see how can a Readme can be useful to maximum power).