Dart Classes and JSON Objects Extraction.
- README.md
There is a correction on the Readme and the addition of the example and explanation of the Video #4

- DART_CLASES and PERSONA
This files are all about classes and how to work with them.

- SPACEX_JSON
Is a folder that contains the launches of SPACE X at July 17th of 2021 and a little example of how to extract objects in the JSON format.