MaterialApp (Video #8)
This is a more complete app, but still being very simple.

If you are asking why is this repositorie, making a commit every single video, is to know, which are the program of each video, and divide every simple part, because creating a project, every video, will consume to much space and compile will be head pain.