Flutter - Image
The file that explains the Widget Image is uploaded.
The Widget Image is useful to use images on the App.
Then, It is need it to create a folder name: assets and there, you will upload all the images that you want in your project (Obviously, those that are from an URL are not going here).
And you will need to upload the pubspec.yaml and uncomment the assets and add manually (I don't know if there can be added automatically) the files inside assets.