Use of Lint Inside the Project.
In this commit is the addition of the package Lint (If you want to find it: https://pub.dev/packages/lint ).

All the files that are deleted is because there are renamed with the following format: lowercase_only.dart

And in the majority of the changes, are principal to put the values as a constant or commenting the default values.