Course
- pubspec.lock & pubspec.yaml
Both are actualized, on the pubspec.yaml is added the http: ^0.13.3 that is used on the flutter_user_list.dart

- flutter_alert_dialog.dart
This one is to exemplify the how an AlertDialog works. Is simple and makes to the student knows which is the way, how to do that the user really needs to press inside the dialog…

- flutter_color_edit.dart
In this case, the main purpose is to see how you can pass information from one screen to the other. Using more than one StatefulWidget and modifying the parameters in both screens.

- flutter_hello_advanced.dart
In this file, the StatefulWidget Saludo, show a text input, where you need to put your name and then, when you pressed the ‘Saluda’ button, the app will give you an Alert Dialog with Hello $name, how are you?

- flutter_hello_beginner.dart
In this case, we are using the showDialog(){…} to represent how an Alert works on the Screen.

- flutter_parameters.dart
In this case the Parameters are passed from _TextScreenState that contains one variable to EditScreen that contents the other variable and viceversa.
_TextScreenState contains the text itself and in the other class (EditScreen) the user change the value and then pass it to the principal page.

- flutter_routes.dart
This app is only to show how an interaction between two or more screens work.

- flutter_user_list.dart
In this file, we are putting in practice how to work with JSON and making http request, the best part is to take practice with the most popular and useful libraries that are on flutter.
We create an User class, that has fullname, username and photoURL and the the principal constructor and a constructor that is .fromJSON(…).